fix(autoindent): disabled global autocommand for autoindent (options.lua)
This is needed if working on several open source projects, which has different indent rules/settings
than my nvim setup. Else, every file which will be saved would change every indent of the file,
which will cause that github will tell, that every line was edited, because it changed the indent on
all lines. So this setting is NECESSARY, if working with other Projects than the own ones.